checks for files before moving during install

+++ install.sh
@@ -1,8 +1,10 @@
 mkdir -p ~/backups
 for f in .asdfrc .bash_profile .bashrc .editrc .gemrc .gitconfig .inputrc .irbrc .pryrc .rspec .tmux.conf .vim .vimrc
 do
-mv ~/$f ~/backups/$f
-ln -s ~/dev/dotfiles/$f ~/$f
+  if [[ -f "$f" ]]; then
+    mv ~/$f ~/backups/$f
+  fi
+  ln -s ~/dev/dotfiles/$f ~/$f
 done
